refactor: updated env config, renamed util functions and minor controller status code updates
