+++ .env.example
@@ -1,6 +1,6 @@
 # DATABASE_URL=postgres://admin:root@localhost:5432
 
-APP_ENV=development
+NODE_ENV=development
 
 SERVER_PORT=3000
