+++ src/modules/auth/auth.controller.ts
@@ -3,7 +3,7 @@ import { Router } from "express";
 import { AuthService } from "./auth.service";
 import { loginRequestBodyDtoSchema } from "./dto/request/login-request-body.dto";
 import { signUpRequestBodyDtoSchema } from "./dto/request/sign-up-request-body.dto";
-import { sendCookies } from "#/shared/utils/sendCookies";
+import { sendCookies } from "#/shared/utils/send-cookies";
 import "express-async-errors";
 
 export const AuthController = Router();
@@ -33,7 +33,7 @@ AuthController.post(
 		sendCookies(res, "refreshToken", refreshToken, 7 * 24 * 60 * 60 * 1000);
 
 		return res
-			.status(201)
+			.status(200)
 			.json({ data: newUser, message: "User signed up successfully" });
 	}
 );
