+++ src/shared/utils/generate-token.ts
