+++ src/shared/utils/send-cookies.ts
@@ -1,10 +1,11 @@
+import { envConfig } from "#/config/env.config";
 import { Response } from "express";
 
 export function sendCookies(res: Response, name: string, data: string, age: number) {
 	res.cookie(name, data, {
 		httpOnly: true,
-		secure: true,
-		sameSite: "none",
+		secure: envConfig["NODE_ENV"] === "production",
+		sameSite: envConfig["NODE_ENV"] === "production" ? "strict" : "none",
 		maxAge: age,
 	});
 }
