+++ src/shared/utils/verify-token.ts
